update code email

+++ QueuePid/queueman/models.py
@@ -5,7 +5,7 @@ class Queueman(models.Model):
     username = models.OneToOneField(User,on_delete=models.CASCADE)
     phone_number = models.CharField(max_length=10)
     line_id = models.CharField(max_length=30)
-    star = models.IntegerField(default=5)
+    star = models.FloatField(default=5)
     credit = models.PositiveIntegerField(default=0)
 
     def __str__(self):
